backend/requirements.txt
anyio==3.6.1
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.0
filelock==3.8.0
google-api-core==2.10.1
google-auth==2.11.0
google-cloud-vision==3.1.2
googleapis-common-protos==1.56.4
grpcio==1.49.0
grpcio-status==1.49.0
h11==0.13.0
httptools==0.5.0
huggingface-hub==0.9.1
idna==3.4
jellyfish==0.9.0
joblib==1.2.0
networkx==2.8.6
nltk==3.7
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
proto-plus==1.22.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyparsing==3.0.9
python-dotenv==0.21.0
PyYAML==6.0
rake-nltk==1.0.6
regex==2022.9.13
requests==2.28.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
segtok==1.5.11
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tabulate==0.8.10
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.22.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websockets==10.3
yake==0.4.8
